updated PreferenceRepository.kt for goals and settings
