--- db/src/main/java/com/bruhascended/db/preferences/ActivityPreferences.kt
@@ -1,8 +1,0 @@
-package com.bruhascended.db.preferences
-
-data class ActivityPreferences(
-    var calories: Int,
-    var distance: Double,
-    var steps: Int,
-    var duration: Long,
-)
