--- db/src/main/java/com/bruhascended/db/preferences/NutritionPreferences.kt
@@ -1,8 +1,0 @@
-package com.bruhascended.db.preferences
-
-data class NutritionPreferences(
-    var calories: Int,
-    var proteins: Double,
-    var fats: Double,
-    var carbs: Double,
-)
